file restructure on models_routers
