+++ database/models_routers/ingredients/ingredientsRouter.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 const ingredient = require('./ingredientsModel');
 
-router.get('/all', (req, res) => {
+router.get('/allIngredients', (req, res) => {
   ingredient.findAllIngrdients()
       .then(ingredients => {
         res.json(ingredients);
@@ -17,11 +17,11 @@ router.get('/:id', (req, res) => {
       if (ingredient) {
         res.json(ingredient);
       } else {
-        res.status(404).json({ message: 'Could not find user with given id.' })
+        res.status(404).json({ message: 'Could not find ingredient with given id.' })
       }
     })
     .catch(err => {
-      res.status(500).json({ message: 'Failed to get user' });
+      res.status(500).json({ message: 'Failed to get ingredient' });
     });
   });
