+++ database/models_routers/users/users.spec.js
@@ -1,0 +1,8 @@
+describe('the server', () => {
+    describe('GET /', () => {
+        it('should run the testing env', () => {
+            expect(process.env.DB_ENV).toBe('testing');
+        })
+    })
+});
+
